git-tfs-id: [http://ros-tfs-px01.picis.com:8080/tfs/emar]$/;C365

+++ WebAPI/API/C_Development/Work/Antoni/Emar.Core/Medications/Model/MedicationRouteDto.cs
@@ -7,7 +7,7 @@
         /// </summary>
         public int Id { get; set; }
 
-        public long SiteId { get; set; }
+        public int SiteId { get; set; }
 
         /// <summary>
         /// Medication route name.
