+++ WebAPI/API/C_Development/Work/Antoni/Emar.Data/Entities/SiteOption.cs
@@ -10,7 +10,7 @@ namespace Emar.Data.Entities
         [Column("id")]
         public int Id { get; set; }
         [Column("site_id", TypeName = "int")]
-        public long SiteId { get; set; }
+        public int SiteId { get; set; }
         [Column("option_id")]
         public int OptionId { get; set; }
         [Required]
@@ -21,6 +21,7 @@ namespace Emar.Data.Entities
         [ForeignKey(nameof(OptionId))]
         [InverseProperty(nameof(Entities.Option.SiteOptions))]
         public virtual Option Option { get; set; }
+
         [ForeignKey(nameof(SiteId))]
         [InverseProperty(nameof(Entities.Site.SiteOptions))]
         public virtual Site Site { get; set; }
